edit: parser함수 수정, tokenizer 관련 함수들 프로토타입 변경
